# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.",
        "title": "eID based AI enabled Government Template Platform V3.0 API",
        "termsOfService": "https://github.com/snykk/go-rest-boilerplate",
        "contact": {
            "name": "eID based AI enabled Government Template Platform V3.0",
            "url": "https://github.com/snykk/go-rest-boilerplate"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/me/eid/activity": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "eID auth/sign түүх + тоо (RP-scoped)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "хуудасны хэмжээ (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "эхлэх байрлал",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDActivityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/me/eid/certificates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "eID гэрчилгээний жагсаалт + тоо",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDCertificatesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/me/eid/devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "eID холбоотой төхөөрөмжүүд",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDDevicesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/me/eid/organizations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Нэвтэрсэн eID хэрэглэгчийн eidmongolia.mn-д бүртгэлтэй, төлөөлж чадах байгууллагуудыг буцаана. eID-ээр нэвтрээгүй хэрэглэгчид хоосон жагсаалт.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Төлөөлдөг байгууллагууд (eID)",
                "responses": {
                    "200": {
                        "description": "Representations",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/template_internal_http_datatransfers_responses.OrgRepresentationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "eID provider error",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Улсын бүртгэлээс (XYP) байгууллагыг регистрийн дугаараар хайж, нэвтэрсэн иргэнийг (eID РД нь тухайн байгууллагын захирал/үүсгэн байгуулагч/хувь эзэмшигч бол) eidmongolia.mn-д төлөөлөл болгон холбоно. Иргэний бүх төлөөлдөг байгууллагыг буцаана.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Байгууллага холбох (eID)",
                "parameters": [
                    {
                        "description": "Байгууллагын регистрийн дугаар",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_datatransfers_requests.EIDOrgRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Representations",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/template_internal_http_datatransfers_responses.OrgRepresentationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid body",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Not authorized to represent this organization",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Organization not found",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/users/me/eid/organizations/{regNo}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Нэвтэрсэн иргэн өөрийн төлөөлдөг байгууллагын (регистрийн дугаараар) холбоосыг цуцлана. Иргэний үлдсэн төлөөлдөг байгууллагыг буцаана.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Байгууллага салгах (eID)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Байгууллагын регистрийн дугаар",
                        "name": "regNo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/template_internal_http_datatransfers_responses.OrgRepresentationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/me/eid/organizations/{regNo}/signers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Нэвтэрсэн иргэний төлөөлдөг байгууллагын гарын үсэг зурах эрхтэй хүмүүс. Иргэн тухайн байгууллагын төлөөлөгч байх ёстой.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Байгууллагын гарын үсэг зурагчид (eID)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Байгууллагын регистрийн дугаар",
                        "name": "regNo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/template_internal_http_datatransfers_responses.OrgSignerResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Нэвтэрсэн иргэний төлөөлдөг байгууллагад өөр eID иргэнийг (регистрийн дугаараар) гарын үсэг зурах эрхтэй болгож нэмнэ.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Гарын үсэг зурагч нэмэх (eID)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Байгууллагын регистрийн дугаар",
                        "name": "regNo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Гарын үсэг зурагчийн РД + үүрэг",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_datatransfers_requests.AddEIDSignerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signers + pending confirmation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.OrgSignersResultResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Нэвтэрсэн иргэний төлөөлдөг байгууллагаас гарын үсэг зурагчийг (регистрийн дугаараар) хасна.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Гарын үсэг зурагч хасах (eID)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Байгууллагын регистрийн дугаар",
                        "name": "regNo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Хасах гарын үсэг зурагчийн РД",
                        "name": "signer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/template_internal_http_datatransfers_responses.OrgSignerResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/me/eid/organizations/{regNo}/signers/resend": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Баталгаажаагүй (PENDING) гарын үсэг зурагч руу eID sign-push баталгаажуулах хүсэлтийг дахин илгээнэ.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Баталгаажуулах хүсэлт дахин илгээх (eID)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Байгууллагын регистрийн дугаар",
                        "name": "regNo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Гарын үсэг зурагчийн РД",
                        "name": "signer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.OrgSignersResultResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "template_internal_http_datatransfers_responses.EIDActivityResponse": {
            "type": "object",
            "properties": {
                "counts": {
                    "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDActivityCounts"
                },
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDActivityItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "template_internal_http_handlers_v1.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "/auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен (access token).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}